TINY-6085 Deleted unused snooker functions/files
