+++ modules/snooker/src/main/ts/ephox/snooker/resize/Sizes.ts
@@ -3,7 +3,7 @@ import { Attr, Css, Element, Height, Node, Width } from '@ephox/sugar';
 import * as TableLookup from '../api/TableLookup';
 import { getSpan } from '../util/CellUtils';
 import * as RuntimeSize from './RuntimeSize';
-import { TableSize } from './Types';
+import { TableSize } from './TableSize';
 
 export interface GenericWidth {
   width: () => number;
