+++ modules/snooker/src/main/ts/ephox/snooker/resize/TableSize.ts
@@ -3,7 +3,19 @@ import { Width, Element } from '@ephox/sugar';
 import * as CellUtils from '../util/CellUtils';
 import * as ColumnSizes from './ColumnSizes';
 import * as Sizes from './Sizes';
-import { TableSize } from './Types';
+import { Warehouse } from '../model/Warehouse';
+import { BarPositions, ColInfo } from './BarPositions';
+
+export interface TableSize {
+  width: () => number;
+  pixelWidth: () => number;
+  getWidths: (warehouse: Warehouse, direction: BarPositions<ColInfo>, tableSize: TableSize) => number[];
+  getCellDelta: (delta: number) => number;
+  singleColumnWidth: (w: number, delta: number) => number[];
+  minCellWidth: () => number;
+  setElementWidth: (cell: Element, amount: number) => void;
+  setTableWidth: (table: Element, newWidths: number[], delta: number) => void;
+}
 
 const percentageSize = function (width: string, element: Element): TableSize {
   const floatWidth = parseFloat(width);
