+++ modules/snooker/src/main/ts/ephox/snooker/util/Util.ts
@@ -1,7 +1,7 @@
-import { Arr, Option } from '@ephox/katamari';
+import { Arr, Option, Fun } from '@ephox/katamari';
 
 // Rename this module, and repeat should be in Arr.
-const repeat = function <T> (repititions: number, f: (idx: number) => T) {
+const repeat = <T> (repititions: number, f: (idx: number) => T) => {
   const r: T[] = [];
   for (let i = 0; i < repititions; i++) {
     r.push(f(i));
@@ -9,17 +9,11 @@ const repeat = function <T> (repititions: number, f: (idx: number) => T) {
   return r;
 };
 
-const range = function (start: number, end: number) {
-  const r: number[] = [];
-  for (let i = start; i < end; i++) {
-    r.push(i);
-  }
-  return r;
-};
+const range = (max: number) => Arr.range(max, Fun.identity);
 
-const unique = function <T> (xs: T[], eq: (a: T, b: T) => boolean) {
+const unique = <T> (xs: T[], eq: (a: T, b: T) => boolean) => {
   const result: T[] = [];
-  Arr.each(xs, function (x, i) {
+  Arr.each(xs, (x, i) => {
     if (i < xs.length - 1 && !eq(x, xs[i + 1])) {
       result.push(x);
     } else if (i === xs.length - 1) {
@@ -29,38 +23,25 @@ const unique = function <T> (xs: T[], eq: (a: T, b: T) => boolean) {
   return result;
 };
 
-const deduce = function (xs: Option<number>[], index: number) {
+const deduce = (xs: Option<number>[], index: number) => {
   if (index < 0 || index >= xs.length - 1) {
     return Option.none<number>();
   }
 
-  const current = xs[index].fold(function () {
+  const current = xs[index].fold(() => {
     const rest = Arr.reverse(xs.slice(0, index));
-    return Arr.findMap(rest, function (a, i) {
-      return a.map(function (aa) {
-        return { value: aa, delta: i + 1 };
-      });
-    });
-  }, function (c) {
-    return Option.some({ value: c, delta: 0 });
-  });
-  const next = xs[index + 1].fold(function () {
-    const rest = xs.slice(index + 1);
-    return Arr.findMap(rest, function (a, i) {
-      return a.map(function (aa) {
-        return { value: aa, delta: i + 1 };
-      });
-    });
-  }, function (n) {
-    return Option.some({ value: n, delta: 1 });
-  });
+    return Arr.findMap(rest, (a, i) => a.map((aa) => ({ value: aa, delta: i + 1 })));
+  }, (c) => Option.some({ value: c, delta: 0 }));
 
-  return current.bind(function (c) {
-    return next.map(function (n) {
-      const extras = n.delta + c.delta;
-      return Math.abs(n.value - c.value) / extras;
-    });
-  });
+  const next = xs[index + 1].fold(() => {
+    const rest = xs.slice(index + 1);
+    return Arr.findMap(rest, (a, i) => a.map((aa) => ({ value: aa, delta: i + 1 })));
+  }, (n) => Option.some({ value: n, delta: 1 }));
+
+  return current.bind((c) => next.map((n) => {
+    const extras = n.delta + c.delta;
+    return Math.abs(n.value - c.value) / extras;
+  }));
 };
 
 export {
