+++ modules/snooker/src/test/ts/browser/RuntimeSizeTest.ts
@@ -11,10 +11,6 @@ UnitTest.test('Runtime Size Test', function () {
     return Math.round(Math.random() * (max - min) + min);
   };
 
-  const getOuterWidth = function (elm: Element) {
-    return Math.round(elm.dom().getBoundingClientRect().width);
-  };
-
   const getOuterHeight = function (elm: Element) {
     return Math.round(elm.dom().getBoundingClientRect().height);
   };
@@ -28,10 +24,6 @@ UnitTest.test('Runtime Size Test', function () {
       total: getSize(table),
       cells: measureCells(getSize, table)
     };
-  };
-
-  const setWidth = function (table: Element, value: string) {
-    Css.set(table, 'width', value);
   };
 
   const setHeight = function (table: Element, value: string) {
@@ -75,58 +67,6 @@ UnitTest.test('Runtime Size Test', function () {
   const randomBorder = function (min: number, max: number, color: string) {
     const n = random(min, max);
     return n > 0 ? n + 'px solid ' + color : '0';
-  };
-
-  const createTableH = function () {
-    const table = Element.fromTag('table');
-    const tbody = Element.fromTag('tbody');
-    const tr = Element.fromTag('tr');
-
-    Attr.set(table, 'border', '1');
-    Attr.set(table, 'cellpadding', random(0, 10).toString());
-    Attr.set(table, 'cellspacing', random(0, 10).toString());
-
-    Css.setAll(table, {
-      'border-collapse': randomValue([ 'collapse', 'separate' ]),
-      'border-left': randomBorder(0, 5, 'red'),
-      'border-right': randomBorder(0, 5, 'red'),
-      'width': randomSize(100, 1000)
-    });
-
-    const cells = Arr.range(random(1, 5), function (_) {
-      const cell = Element.fromTag('td');
-
-      Css.setAll(cell, {
-        'width': randomSize(1, 100),
-        'height': '10px',
-        'padding-left': randomSize(0, 5),
-        'padding-right': randomSize(0, 5),
-        'border-left': randomBorder(0, 5, 'green'),
-        'border-right': randomBorder(0, 5, 'green')
-      });
-
-      const content = Element.fromTag('div');
-
-      Css.setAll(content, {
-        width: randomSize(1, 200),
-        height: '10px'
-      });
-
-      Insert.append(cell, content);
-
-      return cell;
-    });
-
-    Insert.append(table, tbody);
-    Insert.append(tbody, tr);
-
-    Arr.each(cells, function (cell) {
-      Insert.append(tr, cell);
-    });
-
-    Insert.append(Body.body(), table);
-
-    return table;
   };
 
   const createTableV = function () {
@@ -213,6 +153,5 @@ UnitTest.test('Runtime Size Test', function () {
     Arr.each(Arr.range(n, Fun.identity), generator);
   };
 
-  generateTest(testTableSize(createTableH, getOuterWidth, RuntimeSize.getWidth, setWidth), 50);
   generateTest(testTableSize(createTableV, getOuterHeight, RuntimeSize.getHeight, setHeight), 50);
 });
